+++ backend/src/repository/table.rs
@@ -11,7 +11,7 @@ use diesel::prelude::*;
 use diesel::{expression_methods::ExpressionMethods, PgConnection, QueryDsl, RunQueryDsl};
 
 impl Table {
-    pub fn search(conn: &mut PgConnection, dbid: &i32, q: String) -> Result<Vec<Self>, Error> {
+    pub fn search(conn: &mut PgConnection, dbid: &i64, q: String) -> Result<Vec<Self>, Error> {
         tables::table
             .filter(
                 tables::database_id
@@ -25,7 +25,7 @@ impl Table {
 
     pub fn find_matching_by_database_id(
         conn: &mut PgConnection,
-        database_id: i32,
+        database_id: i64,
         prefix: &str,
     ) -> Result<Vec<Self>, Error> {
         tables::table
@@ -41,7 +41,7 @@ impl Table {
     pub fn find_by_name_and_database_id(
         conn: &mut PgConnection,
         name: String,
-        database_id: i32,
+        database_id: i64,
     ) -> Result<Self, Error> {
         tables::table
             .filter(
@@ -56,7 +56,7 @@ impl Table {
     pub fn delete_old_tables(
         conn: &mut PgConnection,
         table_names: &Vec<String>,
-        database_id: i32,
+        database_id: i64,
     ) -> Result<(), Error> {
         conn.transaction::<_, Error, _>(|conn| {
             let table_ids = tables::table
@@ -64,7 +64,7 @@ impl Table {
                     not(tables::name.eq_any(table_names)).and(tables::database_id.eq(database_id)),
                 )
                 .select(tables::id)
-                .load::<i32>(conn)?;
+                .load::<i64>(conn)?;
 
             diesel::delete(columns_::table.filter(columns_::table_id.eq_any(&table_ids)))
                 .execute(conn)?;
@@ -75,7 +75,7 @@ impl Table {
         Ok(())
     }
 
-    pub fn delete_old_primary_keys(conn: &mut PgConnection, database_id: i32) -> Result<(), Error> {
+    pub fn delete_old_primary_keys(conn: &mut PgConnection, database_id: i64) -> Result<(), Error> {
         conn.transaction::<_, Error, _>(|conn| {
             let subquery = tables::table
                 .filter(tables::database_id.eq(database_id))
@@ -93,7 +93,7 @@ impl Table {
 
     pub fn update_primary_keys(
         conn: &mut PgConnection,
-        database_id: i32,
+        database_id: i64,
         primary_keys: Vec<PrimaryKey>,
     ) -> Result<(), Error> {
         conn.transaction::<_, Error, _>(|conn| {
@@ -117,14 +117,14 @@ impl Table {
         Ok(())
     }
 
-    pub fn delete_old_foreign_keys(conn: &mut PgConnection, dbid: i32) -> Result<(), Error> {
+    pub fn delete_old_foreign_keys(conn: &mut PgConnection, dbid: i64) -> Result<(), Error> {
         conn.transaction::<_, Error, _>(|conn| {
             let target_ids = foreign_keys::table
                 .inner_join(columns_::table.on(columns_::id.nullable().eq(foreign_keys::column_id)))
                 .inner_join(tables::table.on(columns_::table_id.eq(tables::id.nullable())))
                 .filter(tables::database_id.eq(dbid))
                 .select(foreign_keys::id)
-                .load::<i32>(conn)?;
+                .load::<i64>(conn)?;
 
             diesel::delete(foreign_keys::table.filter(foreign_keys::id.eq_any(target_ids)))
                 .execute(conn)?;
@@ -135,7 +135,7 @@ impl Table {
     }
     pub fn update_foreign_keys(
         conn: &mut PgConnection,
-        dbid: i32,
+        dbid: i64,
         foreign_keys: Vec<ForeignKey>,
     ) -> Result<(), Error> {
         conn.transaction::<_, Error, _>(|conn| {
@@ -146,14 +146,14 @@ impl Table {
                     .filter(tables::name.eq(fkey.table_name))
                     .filter(columns_::name.eq(fkey.column_name))
                     .select(columns_::id)
-                    .first::<i32>(conn)?;
+                    .first::<i64>(conn)?;
                 let foreign_column_id = columns_::table
                     .inner_join(tables::table.on(tables::id.nullable().eq(columns_::table_id)))
                     .filter(tables::database_id.eq(dbid))
                     .filter(tables::name.eq(fkey.foreign_table_name))
                     .filter(columns_::name.eq(fkey.foreign_column_name))
                     .select(columns_::id)
-                    .first::<i32>(conn)?;
+                    .first::<i64>(conn)?;
                 let fkey_changeset = ForeignKeyChangeset {
                     name: Some(fkey.relation_name),
                     fk_type: Some(FkType::Fk),
